+++ src/TemperatureInCelsiusDecoder.cpp
@@ -7,7 +7,7 @@ using namespace std;
 
 vector <SensorValue*> TemperatureInCelsiusDecoder::decode(const vector<uint8_t> &EncodedMessage){
     vector <SensorValue*> tempValueCelsius;
-    double tempCelsius = EncodedMessage.at(2);
+    uint8_t tempCelsius = EncodedMessage.at(2);
     SensorValue* value = new SensorValue(tempCelsius,"Temperature","Celsius");
         tempValueCelsius.push_back(value);
         return tempValueCelsius;
